Create results page and move results components to component directory, add Suspense boundaries
Moves components to correct directories rather than clog up the routing directory in /app. Also fixes some Suspense boundaries errors cropping up during "npm run build".
